requirements.txt
asgiref==3.6.0
astroid==2.14.1
autopep8==2.0.1
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
dill==0.3.6
Django==4.1.6
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
pep8==1.7.1
Pillow==9.4.0
platformdirs==3.0.0
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.16.1
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
pytz==2023.2
razorpay==1.3.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.7
sqlparse==0.4.3
timedelta==2020.12.3
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.14.1
